+++ HipAndClavicle/.log
@@ -1,0 +1,1 @@
+Attempting to cancel the build...
