+++ HipAndClavicle/Views/Admin/Products.cshtml.css
@@ -1,0 +1,4 @@
+﻿.product-list-view {
+    display:flex;
+    flex-wrap:wrap;
+}
