+++ HipAndClavicle/Views/Shared/DisplayTemplates/Product.cshtml.css
@@ -1,0 +1,30 @@
+﻿
+.product-card {
+    min-width: 180px;
+    max-width: 180px;
+    min-height: 230px;
+    max-height: 230px;
+    margin: 1em;
+    padding: 0;
+}
+
+.color-splotch {
+    min-width: 20px;
+    max-width: 20px;
+    min-height: 20px;
+    max-height: 20px;
+    border-radius: 50%;
+    display: inline-block;
+    margin: 4px;
+}
+
+.card-footer {
+    margin-block:0;
+    min-height: 50px;
+    max-height: 50px;
+}
+
+.card-body {
+    padding-bottom:0;
+    margin:0;
+}
